+++ agent/src/main/java/com/intergral/deep/agent/types/snapshot/WatchResult.java
@@ -22,10 +22,31 @@ package com.intergral.deep.agent.types.snapshot;
  */
 public class WatchResult {
 
+  /**
+   * Watch source is METRIC.
+   *
+   * @see com.intergral.deep.proto.tracepoint.v1.WatchSource
+   */
+  public static final String METRIC = "METRIC";
+
+  /**
+   * Watch source is WATCH.
+   *
+   * @see com.intergral.deep.proto.tracepoint.v1.WatchSource
+   */
+  public static final String WATCH = "WATCH";
+
+  /**
+   * Watch source is LOG.
+   *
+   * @see com.intergral.deep.proto.tracepoint.v1.WatchSource
+   */
+  public static final String LOG = "LOG";
+
   private final String error;
   private final VariableID result;
   private final String expression;
-  private final boolean isMetric;
+  private final String source;
 
 
   /**
@@ -33,13 +54,13 @@ public class WatchResult {
    *
    * @param expression the expression
    * @param error      the error
-   * @param isMetric is this watch generated from a metric expression
+   * @param source     the source of this result
    */
-  public WatchResult(final String expression, final String error, final boolean isMetric) {
+  public WatchResult(final String expression, final String error, final String source) {
     this.expression = expression;
     this.error = error;
     this.result = null;
-    this.isMetric = isMetric;
+    this.source = source;
   }
 
   /**
@@ -47,11 +68,11 @@ public class WatchResult {
    *
    * @param expression the expression
    * @param result     the result
-   * @param isMetric   is this watch generated from a metric expression
+   * @param source     the source of this result
    */
-  public WatchResult(final String expression, final VariableID result, final boolean isMetric) {
+  public WatchResult(final String expression, final VariableID result, final String source) {
     this.expression = expression;
-    this.isMetric = isMetric;
+    this.source = source;
     this.error = null;
     this.result = result;
   }
@@ -72,7 +93,7 @@ public class WatchResult {
     return this.expression;
   }
 
-  public boolean isMetric() {
-    return isMetric;
+  public String getSource() {
+    return source;
   }
 }