+++ agent/src/test/java/com/intergral/deep/agent/tracepoint/handler/FrameCollectorTest.java
@@ -51,26 +51,26 @@ class FrameCollectorTest {
   @Test
   void evaluateWatchers() throws Throwable {
     Mockito.when(evaluator.evaluateExpression(Mockito.anyString(), Mockito.anyMap())).thenReturn("some result");
-    final IExpressionResult someExpression = frameCollector.evaluateWatchExpression("some expression", false);
+    final IExpressionResult someExpression = frameCollector.evaluateWatchExpression("some expression", "test");
     assertEquals("some expression", someExpression.result().expression());
     assertEquals(1, someExpression.variables().size());
     final Variable variable = someExpression.variables().get("1");
     assertEquals("some result", variable.getValString());
     assertFalse(someExpression.isError());
     assertEquals(Double.NaN, someExpression.numberValue());
-    assertFalse(someExpression.result().isMetric());
+    assertEquals("test", someExpression.result().getSource());
   }
 
   @Test
   void evaluateWatchers_error() throws Throwable {
     Mockito.when(evaluator.evaluateExpression(Mockito.anyString(), Mockito.anyMap())).thenThrow(new RuntimeException("Test exception"));
-    final IExpressionResult someExpression = frameCollector.evaluateWatchExpression("some expression", false);
+    final IExpressionResult someExpression = frameCollector.evaluateWatchExpression("some expression", "test");
     assertEquals("some expression", someExpression.result().expression());
     assertEquals(0, someExpression.variables().size());
     assertEquals("java.lang.RuntimeException: Test exception", someExpression.result().error());
     assertTrue(someExpression.isError());
     assertEquals(Double.NaN, someExpression.numberValue());
-    assertFalse(someExpression.result().isMetric());
+    assertEquals("test", someExpression.result().getSource());
   }
 
   @Test